Bug fix
Finally found and corrected the bug that added mal-formed XML to the
template field of the aiml table from the upload script!

I also added "Pretty Print" to the download script, to make the
downloaded AIML filed easier to read, AND updated the bot_test.aiml file
to the "pretty Printed" version.