Bug fix, added PHP version to the debug file
It seems that the PHP manual and PHP itself have different descriptions
for how filter_input_array() work, so my last "fix" did not work as
expected. I think I've got a handle on what was going wrong, though.
Verified to work in Windows and CentOS with PHP versions 5.5.12 and
5.3.3 respectively, but it should also work for PHP 5.2+ as well (I no
longer have a server that uses PHP 5.2).
Also added reporting of the server's PHP version to the debug file, so
that I can see what I'm dealing with when I read debug files.