Bug Fixes for issues #186 , #185 and (I HOPE) #163
Fixed several bugs in the upload script, mainly dealing with how the
error messages for invalid AIML files are processed: Here's the list:
1.) Fixed a bug where the total files with errors was always shown as
zero
2.) Fixed a bug where the filename was being incorrectly reported by
removing the (unnecessary) filename reference altogether
3.) Added filtering to the error messages to remove unnecessary
information, making the messages easier to read

Fixed several bugs with the install script (issue #185) where the log,
debug and config folders were incorrectly being reported as not
writable.
Added a feature to the core code to reset the chatbot when the user
enters ":reset bot" into the conversation. This is easier to use and
more intuitive than the original "clear properties" (which still works,
BTW).
Added more error logging to various files in order to facilitate
troubleshooting.
Minor formatting change to the admin CSS