Swap PERSON tag functionality, other fixes
After a long discussion with @tringate and testing of functionality on
Pandorabots, it was decided that Program O was not properly handling the
`<person>` and `<person2>` tags, so we addressed this issue and swapped
the functionality to work the same as Pandorabots.
Additionally, some code was added to prevent the `runDebug()` function
from executing if the debug level is set to 0 (no debugging). This
should provide a significant boost to performance when debugging has
been disabled.
@tringate also pointed out a minor typo in some of the code, and that
has been fixed as well.