New Requirements Checklist in the Install
A new "Requirements Checklist" has been added to the install script to
show botmasters what PHP extensions are required and their current
state, as well as whether the current PHP version is sufficient to run
Program O. It also shows the current state of the three "recommended"
PHP extensions (cURL, zip and mbstring). The install script will provide
warnings about the missing recommended extensions, but will allow the
user to continue if all other requirements are met. If the PHP version
is too low, or there are no suitable PDO extensions available, then an
error message is displayed, and installation is halted.