Update to improve error message.
I was trying to install, had folders without correct security. One was the debug folder. I assumed it was located with the other two, I couldn't locate it. So the error now tells there user where the folder is located so they know exactly where to go in order to correct the folder security.