Minor format change, fix bug in spellchecker
Reduced the width and font size in the div that contains the script
version information when the logo is hidden, and moved the "Toggle the
Logo" link to the bottom of the list of links. Also corrected a bug in
the spellchecker where a session variable was being improperly accessed.