Code Refactor, new <date> tag attributes
Bug Fixes
The following bugs were found and fixed:
1.  A bug in Windows with certain locales (Hungarian, for example) have
no UTF-8 support, so
were triggering errors in the code that parses the AIML `<date>` tag.
This has been fixed.
2.  While testing the fix for the `<date>` tag it was discovered that
the downloads page
was adding files to the Zip archive to be downloaded that didn't belong
to the selected
chatbot. this isn't noticable when there is only one bot, but with
multiple chatbots
it created an issue. This has also been fixed.

Code Refactoring:
Please see CHANGELOG.md for details

New <date> Tag Attributes:
There is a lot of information available in the CHANGELOG for this, so
please refer to that document for details.