Bug Fixes, minor code refactor
Fixed a bug in the script that was causing certain form variables to
show up as empty (issue #346)
Fixed several typos in various scripts that weren't necessarily causing
issues, but still needed to be corrected.
Refactored the CSS code for some of the admin pages to make certain data
tables present more correctly.