bug fixes, code cleanup, add `<thatstar>` parsing
As a result of issue #354  it was decided to completely refactor the
function that sets wildcard values, addressing both bugs in properly
parsing `<thatstar>` tags, __AND__ the fact that there was no support at
all for `<topicstar>` tags.

Additionally, cleaned up some more orphaned debugging code, either
commenting the code, or removing it altogether.

Regarding the previous lack of support for the `<topicstar>` tag, a new
function was put in place to handle this.