Custom debugging support built into HTML GUI
In an effort to help botmasters with troubleshooting their chatbots,
custom debugging code was added to the HTML GUI. In order to take
advantage of this new functionality, a constant in the file
config/global_config.php (ERROR_DEBUGGING)  must be set to TRUE. Once
that constant has been set to true, a new DIV element is created on page
load that loads the most recent debug file for the `convo_id` of
'DEBUG', allowing botmasters to view both the bot's responses _AND_ the
contents of the debug file on the same page. The debug file viewer
(chatbot/debug/index.php) is still available, and it continues to be
more feature-rich, but for those who prefers to save the time and effort
of flipping between two or more pages to get essentially the same data,
this is now here for you.